add shape interface + test

+++ part_4_oop/squares.go
@@ -4,17 +4,17 @@ package part4
 type Square struct {
 	X      int
 	Y      int
-	Length int
+	Length float64
 }
 
 //NewSquare Square factory
-func NewSquare(x int, y int, length int) *Square {
+func NewSquare(x int, y int, length float64) *Square {
 	newSquare := Square{x, y, length}
 	return &newSquare
 }
 
 //Area return square area
-func (s Square) Area() int {
+func (s Square) Area() float64 {
 	return s.Length * s.Length
 }
